requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
pydantic>=2.6.0
slowapi>=0.1.9
structlog>=24.1.0
httpx>=0.27.0
pytest>=8.1.1
pytest-asyncio>=0.23.6
python-multipart>=0.0.9
